+++ app/actions/change_profile.py
@@ -1,13 +1,15 @@
+''' "profile" command functionalities '''
 from app.config.configuration import Configuration
 from app.config.profile import Profile
 from app.interface import Interface
+from app.message_priority import MessagePriority
 
 def _display_profile(profile: Profile, user_interface: Interface) -> None:
     """Display throught the interface the profile's presentation including metadatas
     @param profile: Profile
     @param user_interface: Interface
     """
-    user_interface.request_output_to_user(f"Profile: {profile.name}")
+    user_interface.request_output_to_user(f"Profile: {profile.name}", MessagePriority.INFO)
 
     unique_tags = set()
     for audio_metadata in profile.audio_metadatas:
@@ -15,17 +17,22 @@ def _display_profile(profile: Profile, user_interface: Interface) -> None:
     unique_tags = list(unique_tags)
     unique_tags.sort()
     tags_list = ", ".join(unique_tags)
-    user_interface.request_output_to_user(f"\nTags found:\n {tags_list}")
+    user_interface.request_output_to_user(f"\nTags found:\n {tags_list}", MessagePriority.INFO)
 
-    user_interface.request_output_to_user("\nAudio's list:")
-    for audio_metadata in profile.audio_metadatas:
+    AUDIOS_TO_SHOW = 5
+    TOTAL_AUDIOS_NUMBER = len(profile.audio_metadatas)
+    user_interface.request_output_to_user(f"\nAudio list ({AUDIOS_TO_SHOW} first audios out of a total of {TOTAL_AUDIOS_NUMBER} audios):", MessagePriority.INFO)
+    for audio_i, audio_metadata in enumerate(profile.audio_metadatas):
+        if audio_i == AUDIOS_TO_SHOW:
+            break
         track_info = f"{audio_metadata.name}"
         track_info += f"by {audio_metadata.author}" if audio_metadata.author is not None else ''
         track_info += ', tags: '
         track_info += f"{audio_metadata.tags}" if audio_metadata.tags else 'no'
         track_info += ', source: ' + 'yes' if audio_metadata.source != '' else 'no'
         user_interface.request_output_to_user(
-            f"- {track_info}"
+            f"- {track_info}",
+            MessagePriority.INFO
         )
 
 def request_profile(args, configuration: Configuration, user_interface: Interface) -> Profile:
@@ -41,7 +48,8 @@ def request_profile(args, configuration: Configuration, user_interface: Interfac
     if len(args) == 1:
         if configuration.profile is None:
             user_interface.request_output_to_user(
-                "There is no profile set. Please set a profile."
+                "There is no profile set. Please set a profile.",
+                MessagePriority.WARNING
             )
             return None
         _display_profile(configuration.profile, user_interface)
@@ -49,12 +57,12 @@ def request_profile(args, configuration: Configuration, user_interface: Interfac
 
     profile_name = str(args[1]) # TODO is the profile name just a string without space allowed or can it have space?
     if profile_name not in configuration.get_profiles():
-        user_interface.request_output_to_user(f"Warning: Unknown profile: \"{profile_name}\"")
+        user_interface.request_output_to_user(f"Warning: Unknown profile: \"{profile_name}\"", MessagePriority.WARNING)
         return configuration.profile
 
     configuration.profile = Profile(name=profile_name)
     configuration.fill_profile_with_metadata(user_interface)
     n_metadatas = len(configuration.profile.audio_metadatas)
-    user_interface.request_output_to_user(f"Info: Profile updated: \"{n_metadatas}\" lines of metadatas found.")
+    user_interface.request_output_to_user(f"Info: Profile updated: \"{n_metadatas}\" lines of metadatas found.", MessagePriority.INFO)
 
     return configuration.profile