+++ app/actions/play.py
@@ -1,5 +1,8 @@
+''' "Play" command functionnalities '''
+
 from app.cannot_find_a_match_error import CannotFindAMatchError
 from app.interface import Interface
+from app.message_priority import MessagePriority
 from app.search import match_string_among_strings
 from audio.audio_player import AudioPlayer
 
@@ -12,28 +15,38 @@ def play(args: list, player: AudioPlayer, user_interface: Interface) -> None:
             second_argument = args[1]
             audio_index = int(second_argument)
         except ValueError:
-            user_interface.request_output_to_user(f"Warning: Unexpected argument provided: \"{second_argument}\", it was expected to be the index of the audio in the playlist (integer).")
+            user_interface.request_output_to_user(f"Warning: Unexpected argument provided: \
+             \"{second_argument}\", it was expected to be the index of the audio in the playlist \
+             (integer).", MessagePriority.WARNING)
             return
         if not player.play_audio_at_index(audio_index):
-            user_interface.request_output_to_user(f"Warning: Cannot find the \"{audio_index}\" index in the playlist.")
+            user_interface.request_output_to_user(f"Warning: Cannot find the \"{audio_index}\" \
+             index in the playlist.", MessagePriority.WARNING)
             return
     if len(args) > 2:
         try:
             audio_name_to_play = " ".join(args[1:])
         except ValueError:
-            user_interface.request_output_to_user(f"Warning: Unexpected name provided: \"{args[1:]}\", it was expected to be a audio name (string).")
+            user_interface.request_output_to_user(f"Warning: Unexpected name provided: \
+             \"{args[1:]}\", it was expected to be a audio name (string).", MessagePriority.WARNING)
             return
         try:
-            audio_to_play_index = match_string_among_strings(audio_name_to_play, player.playlist.names())
+            audio_to_play_index = match_string_among_strings(
+                audio_name_to_play,
+                player.playlist.names()
+            )
         except CannotFindAMatchError:
-            user_interface.request_output_to_user(f"Warning: Cannot find a matching audio with \"{audio_name_to_play}\".")
+            user_interface.request_output_to_user(f"Warning: Cannot find a matching audio with \
+             \"{audio_name_to_play}\".", MessagePriority.WARNING)
             return
-        else:
-            if not player.play_audio_at_index(audio_to_play_index):
-                user_interface.request_output_to_user(f"Warning: Audio match found but cannot find its \"{audio_to_play_index}\" index in the playlist.")
-            return
+        if not player.play_audio_at_index(audio_to_play_index):
+            user_interface.request_output_to_user(f"Warning: Audio match found but cannot find \
+            its \"{audio_to_play_index}\" index in the playlist.", MessagePriority.WARNING)
+        return
     maybe_played_audio = player.get_playing_audio()
     if maybe_played_audio is None:
-        user_interface.request_output_to_user("Info: Playing..")
+        user_interface.request_output_to_user("Info: Playing..", MessagePriority.INFO)
     else:
-        user_interface.request_output_to_user(f"Info: Playing \"{maybe_played_audio.name_without_extension}\".")
+        user_interface.request_output_to_user(
+            f"Info: Playing \"{maybe_played_audio.name_without_extension}\".", MessagePriority.INFO
+        )