+++ app/actions/volume.py
@@ -1,4 +1,5 @@
 from app.interface import Interface
+from app.message_priority import MessagePriority
 from audio.audio_player import AudioPlayer
 
 def request_volume(args: list, player: AudioPlayer, user_interface: Interface):
@@ -8,13 +9,13 @@ def request_volume(args: list, player: AudioPlayer, user_interface: Interface):
     @param user_interface: Interface
     """
     if len(args) == 1:
-        user_interface.request_output_to_user(f"Info: Audio's volume is {player.get_volume()}%.")
+        user_interface.request_output_to_user(f"Info: Audio's volume is {player.get_volume()}%.", MessagePriority.INFO)
     elif len(args) == 2:
         volume_integer = 0
         try:
             volume_integer = int(args[1])
         except ValueError:
-            user_interface.request_output_to_user(f"Error: the second argument \"{args[1]}\" is not an integer. So, the audio's volume cannot be changed.")
+            user_interface.request_output_to_user(f"Warning: the second argument \"{args[1]}\" is not an integer. So the audio's volume cannot be changed.", MessagePriority.WARNING)
             return
         player.set_volume(volume_integer)
-        user_interface.request_output_to_user("Info: Updating audio's volume")
+        user_interface.request_output_to_user("Info: Updating audio's volume", MessagePriority.INFO)