+++ app/cannot_find_a_match_error.py
@@ -1,2 +1,5 @@
+''' Exception module '''
+
 class CannotFindAMatchError(Exception):
+    ''' Not maching has been found exception '''
     pass