+++ app/interface.py
@@ -1,7 +1,10 @@
+''' App's interface module : getting user's inputs and displaying messages '''
+
 import os
 import platform
 from threading import Lock
 
+from app.message_priority import MessagePriority
 from audio.audio import Audio
 from audio.audio_player import AudioPlayer
 
@@ -11,6 +14,9 @@ class Interface:
     def __init__(self):
         self.stdin_lock = Lock()
         self.stdout_lock = Lock()
+        self.messages_by_priority = [ [] for priority in MessagePriority ]
+        self.priority_locks = { priority:Lock() for priority in MessagePriority }
+        self.msg_prio_mutenesses = { priority:False for priority in MessagePriority }
 
     @staticmethod
     def _clean_terminal() -> None:
@@ -93,13 +99,27 @@ class Interface:
         self.free_user_input()
         return user_input
 
-    def request_output_to_user(self, output: str):
-        """Place to interface's next requested action to user_interface.request_output_to_user the
+    def request_output_to_user(self, output: str, priority: MessagePriority):
+        """Place to interface's next requested action to send the
          <output>
-        @param output: str the message to user_interface.request_output_to_user to the user
+        @param output: str the message to send to the user
+        @param priority: MessagePriority
         """
+        # Store the message by priority
+        with self.priority_locks[priority]:
+            self.messages_by_priority[priority].append(output)
+        # Display messages by priority and ignore messages with muted priority
         with self.stdout_lock:
-            print(output)
+            for msg_priority in MessagePriority:
+                with self.priority_locks[msg_priority]:
+                    if len(self.messages_by_priority[msg_priority]) == 0:
+                        continue
+                    message = self.messages_by_priority[msg_priority].pop(0)
+                    if self.msg_prio_mutenesses[msg_priority]:
+                        # we discard messages in a muted priority/level to not display them later
+                        # when they would not be relevant anymore.
+                        continue
+                    print(message)
 
     def update_app_display(self, player: AudioPlayer):
         """Refresh the terminal to display current status information
@@ -107,4 +127,26 @@ class Interface:
         """
         with self.stdout_lock:
             self._clean_terminal()
-        self.request_output_to_user(Interface.status_information_str(player))
+        self.request_output_to_user(Interface.status_information_str(player), MessagePriority.INFO)
+
+    def mute_message(self, priority: MessagePriority):
+        ''' Mute the messages from the given priority '''
+        with self.priority_locks[priority]:
+            self.msg_prio_mutenesses[priority] = True
+
+    def unmute_message(self, priority: MessagePriority):
+        ''' Unmute the messages from the given priority '''
+        with self.priority_locks[priority]:
+            self.msg_prio_mutenesses[priority] = False
+
+    def muteness_status(self):
+        ''' Display which priorities are muted or not '''
+        muteness_status: str = "Muteness status:"
+        for priority in MessagePriority:
+            status = "NOT muted."
+            if self.msg_prio_mutenesses[priority]:
+                status = "muted."
+            muteness_status += "\n"
+            muteness_status += f"\tThe {priority} priority is {status}"
+
+        return muteness_status