+++ app/termination.py
@@ -1,5 +1,8 @@
+''' App's shutdown process '''
+
 from app.config.configuration import Configuration
 from app.interface import Interface
+from app.message_priority import MessagePriority
 from audio.audio_player import AudioPlayer
 from audio_import.audio_loader import flush_playlist_cache
 
@@ -16,6 +19,7 @@ def clean_app_termination(
     if not configuration.is_audio_cache_persistant():
         dir_path = configuration.get_audios_directory_path()
         user_interface.request_output_to_user(
-            f"Info: Erasing saved audios from the \"{dir_path}\" directory."
+            f"Info: Erasing saved audios from the \"{dir_path}\" directory.",
+            MessagePriority.INFO
         )
         flush_playlist_cache(player.playlist, configuration)