+++ CHANGELOG.md
@@ -4,7 +4,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## [Version 1.3](https://github.com/jliottard/music-list-player/releases/tag/v1.3.0) - 2025-08-13
+## Version [v1.4.0] - unreleased
+
+### Added
+
+- Add the mute and unmute commands based by the message priority;
+- Automatically mute INFO and WARNING message when displaying lyrics;
+- Display messages by priority;
+- Message mutability.
+
+### Fixed
+
+- Increase pylint's score;
+- Reduce profile 'audio list' displayed;
+- Import missing MessagePriority class.
+
+## Version [v1.3.0](https://github.com/jliottard/music-list-player/commit/ba0e93286d431de4ca72b58baf74f8ebf3d17963) - 2025-08-13
 
 ### Added
 
@@ -16,7 +31,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Fixed
 
-- Replace the unmaintained youtube-search 3rd party library for pytubefix search.
+- Replace the unmaintained youtube-search third party library for pytubefix search.
 
 ## [Version 1.2](https://github.com/jliottard/music-list-player/releases/tag/v1.2.0) - 2024-11-17